Extract `subspace-test-client` from `subspace-test-service`
